+++ src/RestauranteFastFood.java
@@ -1,33 +1,33 @@
 public class RestauranteFastFood extends Restaurante {
-    private float numMedioDiarioClientesDriveThru;
-    private float valMedioFaturacaoClienteDriveThru;
+    private double numMedioDiarioClientesDriveThru;
+    private double valMedioFaturacaoClienteDriveThru;
 
     public RestauranteFastFood(){}
 
-    public RestauranteFastFood(String nome, String distrito, GPS localizacao, int numEmpregadosMesa, float numMedClientesDiario, float custoEmpregadosMesa, float custoSalarioMedioAnual, float valMedioFaturacaoMesaDia, int numDiasFuncionamentoAno, int numMesasInteriores, float numMedioDiarioClientesDriveThru, float valMedioFaturacaoClienteDriveThru) {
+    public RestauranteFastFood(String nome, String distrito, GPS localizacao, int numEmpregadosMesa, double numMedClientesDiario, double custoEmpregadosMesa, double custoSalarioMedioAnual, double valMedioFaturacaoMesaDia, int numDiasFuncionamentoAno, int numMesasInteriores, double numMedioDiarioClientesDriveThru, double valMedioFaturacaoClienteDriveThru) {
         super("Restaurante Fast-food", nome, distrito, localizacao, numEmpregadosMesa, numMedClientesDiario, custoEmpregadosMesa, custoSalarioMedioAnual, valMedioFaturacaoMesaDia, numDiasFuncionamentoAno, numMesasInteriores);
         this.numMedioDiarioClientesDriveThru = numMedioDiarioClientesDriveThru;
         this.valMedioFaturacaoClienteDriveThru = valMedioFaturacaoClienteDriveThru;
     }
 
-    public float getNumMedioDiarioClientesDriveThru() {
+    public double getNumMedioDiarioClientesDriveThru() {
         return numMedioDiarioClientesDriveThru;
     }
 
-    public void setNumMedioDiarioClientesDriveThru(float numMedioDiarioClientesDriveThru) {
+    public void setNumMedioDiarioClientesDriveThru(double numMedioDiarioClientesDriveThru) {
         this.numMedioDiarioClientesDriveThru = numMedioDiarioClientesDriveThru;
     }
 
-    public float getValMedioFaturacaoClienteDriveThru() {
+    public double getValMedioFaturacaoClienteDriveThru() {
         return valMedioFaturacaoClienteDriveThru;
     }
 
-    public void setValMedioFaturacaoClienteDriveThru(float valMedioFaturacaoClienteDriveThru) {
+    public void setValMedioFaturacaoClienteDriveThru(double valMedioFaturacaoClienteDriveThru) {
         this.valMedioFaturacaoClienteDriveThru = valMedioFaturacaoClienteDriveThru;
     }
 
     @Override
-    public float defReceitaAnual() {
+    public double defReceitaAnual() {
         return (numMesasInteriores * valMedioFaturacaoMesaDia + numMedioDiarioClientesDriveThru * valMedioFaturacaoClienteDriveThru) * numDiasFuncionamentoAno;
     }
